refactor: move Report implementations to output.rs and fix clippy warnings
